--- cfg/params.yaml
@@ -1,71 +1,0 @@
-#############################################################
-#                                                           #
-# Copyright (c) 2022, University of California, Los Angeles #
-#                                                           #
-# Authors: Kenny J. Chen, Brett T. Lopez                    #
-# Contact: kennyjchen@ucla.edu, btlopez@ucla.edu            #
-#                                                           #
-#############################################################
-
-dlo:
-
-  odomNode:
-
-    initialPose:
-      use: false
-      position:
-        x: 0.0
-        y: 0.0
-        z: 0.0
-      orientation:
-        w: 1.0
-        x: 0.0
-        y: 0.0
-        z: 0.0
-
-    preprocessing:
-      cropBoxFilter:
-        use: true
-        size: 1.0
-      voxelFilter:
-        scan:
-          use: true
-          res: 0.25
-        submap:
-          use: true
-          res: 0.5
-
-    keyframe:
-      threshD: 5.0
-      threshR: 45.0
-
-    submap:
-      keyframe:
-        knn: 10
-        kcv: 10
-        kcc: 10
-
-    imu:
-      calibTime: 3
-      bufferSize: 2000
-
-    gicp:
-      minNumPoints: 10
-      s2s:
-        kCorrespondences: 10
-        maxCorrespondenceDistance: 1.0
-        maxIterations: 32
-        transformationEpsilon: 0.01
-        euclideanFitnessEpsilon: 0.01
-        ransac:
-          iterations: 5
-          outlierRejectionThresh: 1.0
-      s2m:
-        kCorrespondences: 20
-        maxCorrespondenceDistance: 0.5
-        maxIterations: 32
-        transformationEpsilon: 0.01
-        euclideanFitnessEpsilon: 0.01
-        ransac:
-          iterations: 5
-          outlierRejectionThresh: 1.0
