--- include/dlo/dlo.h
@@ -1,62 +1,0 @@
-/************************************************************
- *
- * Copyright (c) 2022, University of California, Los Angeles
- *
- * Authors: Kenny J. Chen, Brett T. Lopez
- * Contact: kennyjchen@ucla.edu, btlopez@ucla.edu
- *
- ***********************************************************/
-
-#include <atomic>
-#include <fstream>
-#include <iomanip>
-#include <ios>
-#include <iostream>
-#include <mutex>
-#include <signal.h>
-#include <sstream>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string>
-#include <sys/times.h>
-#include <sys/vtimes.h>
-#include <thread>
-
-#ifdef HAS_CPUID
-#include <cpuid.h>
-#endif
-
-#include <ros/ros.h>
-#include <boost/circular_buffer.hpp>
-#include <boost/algorithm/string.hpp>
-
-#include <pcl/filters/crop_box.h>
-#include <pcl/filters/voxel_grid.h>
-#include <pcl/io/pcd_io.h>
-#include <pcl/surface/concave_hull.h>
-#include <pcl/surface/convex_hull.h>
-#include <pcl_conversions/pcl_conversions.h>
-#include <pcl_ros/impl/transforms.hpp>
-#include <pcl_ros/point_cloud.h>
-#include <pcl_ros/transforms.h>
-#include <tf2_ros/transform_broadcaster.h>
-
-#include <geometry_msgs/PoseStamped.h>
-#include <nav_msgs/Odometry.h>
-#include <sensor_msgs/CameraInfo.h>
-#include <sensor_msgs/Image.h>
-#include <sensor_msgs/Imu.h>
-#include <sensor_msgs/PointCloud2.h>
-
-#include <direct_lidar_odometry/save_pcd.h>
-#include <direct_lidar_odometry/save_traj.h>
-#include <nano_gicp/nano_gicp.hpp>
-
-typedef pcl::PointXYZI PointType;
-
-namespace dlo {
-
-  class OdomNode;
-  class MapNode;
-
-}
