+++ include/point_type_nano_gicp.hpp
@@ -1,0 +1,10 @@
+#ifndef POINT_TYPE_NANO_GICP_HPP
+#define POINT_TYPE_NANO_GICP_HPP
+
+
+#include <pcl/point_cloud.h>
+#include <pcl/point_types.h>
+typedef pcl::PointXYZI PointType;
+
+
+#endif
