--- srv/save_pcd.srv
@@ -1,4 +1,0 @@
-float32 leaf_size
-string save_path
----
-bool success
