--- srv/save_traj.srv
@@ -1,3 +1,0 @@
-string save_path
----
-bool success
